Extracted base class
