+++ sead.query.test/IntegrationTests/SeadControllers/QueryBuilderTests.cs
@@ -16,31 +16,8 @@ namespace IntegrationTests.Sead
 {
 
     [Collection("SeadJsonFacetContextFixture")]
-    public class QueryBuilderTests
+    public class QueryBuilderTests(SeadJsonFacetContextFixture facetContext) : IntegrationTestBase(facetContext)
     {
-        public JsonFacetContextFixture FacetContextFixture { get; }
-        public DependencyService DependencyService { get; }
-        public IContainer Container { get; private set; }
-        public IResultSqlCompilerLocator SqlCompilerLocator { get; private set; }
-        public IQuerySetupBuilder QuerySetupBuilder { get; private set; }
-        public IRepositoryRegistry Registry { get; private set; }
-        
-        public QueryBuilderTests(SeadJsonFacetContextFixture facetContextFixture)
-        {
-            FacetContextFixture = facetContextFixture;
-            DependencyService = new DependencyService(FacetContextFixture) { Options = SettingFactory.GetSettings() };
-            var builder = new ContainerBuilder();
-            builder.RegisterModule(DependencyService);
-            Container = builder.Build();
-            QuerySetupBuilder = Container.Resolve<IQuerySetupBuilder>();
-            SqlCompilerLocator = Container.Resolve<IResultSqlCompilerLocator>();
-            Registry = Container.Resolve<IRepositoryRegistry>();
-        }
-        public FacetsConfig2 FakeFacetsConfig(string uri)
-            => new MockFacetsConfigFactory(Registry.Facets).Create(uri);
-
-        public virtual ResultConfig FakeResultConfig(string facetCode, string specificationKey, string viewTypeId)
-            => ResultConfigFactory.Create(Registry.Facets.GetByCode(facetCode), Registry.Results.GetByKey(specificationKey), viewTypeId);
 
         [Theory]
         [InlineData("species:species")]
